requirements.txt
numpy
torchvision
pytorch-lightning
FastAPI
uvicorn
opencv-python

